+++ http/menus.http
@@ -15,6 +15,6 @@ Content-Type: application/json
 }
 
 ###
-GET {{host}}/api/menus
+GET localhost:8080/api/menus
 
 ###
