+++ src/main/java/kitchenpos/application/MenuGroupService.java
@@ -2,8 +2,8 @@ package kitchenpos.application;
 
 import java.util.List;
 import kitchenpos.application.dto.MenuGroupCreateDto;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.MenuGroupRepository;
+import kitchenpos.domain.menu.MenuGroup;
+import kitchenpos.domain.menu.MenuGroupRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
