+++ src/main/java/kitchenpos/application/dto/ProductDto.java
@@ -1,7 +1,7 @@
 package kitchenpos.application.dto;
 
 import java.math.BigDecimal;
-import kitchenpos.domain.Product;
+import kitchenpos.domain.product.Product;
 
 public class ProductDto {
