+++ src/main/java/kitchenpos/domain/order/OrderLineItem.java
@@ -1,10 +1,11 @@
-package kitchenpos.domain;
+package kitchenpos.domain.order;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.ManyToOne;
+import kitchenpos.domain.menu.Menu;
 
 @Entity
 public class OrderLineItem {
