+++ src/main/java/kitchenpos/domain/table/TableGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.table;
 
 import java.time.LocalDateTime;
 import java.util.List;
