+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -6,8 +6,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.application.dto.ProductCreateDto;
-import kitchenpos.domain.Product;
 import kitchenpos.domain.exception.ProductException.NoPriceException;
+import kitchenpos.domain.product.Product;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullSource;
